requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
antlr4-python3-runtime==4.9.3
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
Brotli==1.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
comm==0.2.2
contourpy==1.3.2
cssselect2==0.8.0
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
einops==0.7.0
executing==2.2.0
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
grpcio==1.73.1
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lightning-utilities==0.14.3
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
memory-profiler==0.61.0
mpmath==1.3.0
multidict==6.6.3
nest-asyncio==1.6.0
networkx==3.5
numpy==1.26.4
omegaconf==2.3.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.28.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.2
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytorch-tabnet==4.1.0
pytorch-tabular==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
requests==2.32.4
rich==14.0.0
scikit-learn==1.6.1
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
sklearn-compat==0.1.3
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
torch==2.7.1
torchmetrics==1.5.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
weasyprint==66.0
webencodings==0.5.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
xgboost==3.0.2
yarl==1.20.1
zopfli==0.2.3.post1
